# internal/adapters/delivery/http/handlers/dto/request/user_dto.go
package request

type UserDTO struct {
	ID       string `json:"id,omitempty"`
	Name     string `json:"name"`
	Email    string `json:"email"`
	Password string `json:"password"`
	Admin    bool   `json:"admin,omitempty"`
}